requirements.txt
omegaconf>=2.0.0
tokenizers>=0.9.0
tqdm>=4.48.0
pytorch-lightning==1.0.0
wandb>=0.10.5
sacrebleu>=1.4.14
datasets>=1.1.2
